Fremont Bridge Bicycle Traffic Analysis
Thanks to Jake Vanderplas for putting the layout of this analysis together. I am happy to have learned the majority of my Data Science Knowledge from your priceless book.  
Also, I am happy to have personally contributed to this dataset, as riding a bicycle over the Fremont Bridge was part of my daily commute for 2 years.